+++ hawtio-system/src/main/java/io/hawt/web/proxy/ProxyDetails.java
@@ -147,19 +147,19 @@ public class ProxyDetails implements ProxyAddress {
         }
     }
 
-    public boolean isAllowed(Set<String> whitelist) {
-        if (whitelist.contains("*")) {
+    public boolean isAllowed(Set<String> allowlist) {
+        if (allowlist.contains("*")) {
             return true;
         }
         // host may contain port number! (e.g. "localhost:9000")
-        return whitelist.contains(host.split(":")[0]);
+        return allowlist.contains(host.split(":")[0]);
     }
 
-    public boolean isAllowed(List<Pattern> regexWhitelist) {
+    public boolean isAllowed(List<Pattern> regexAllowlist) {
         // host may contain port number! (e.g. "localhost:9000")
         String hostWithoutPort = host.split(":")[0];
 
-        for (Pattern pattern : regexWhitelist) {
+        for (Pattern pattern : regexAllowlist) {
             if (pattern.matcher(hostWithoutPort).matches()) {
                 return true;
             }